Change Class Name to Jarvis
