--- src/main/java/duke/ui/DialogBox.java
@@ -1,84 +1,0 @@
-package duke.ui;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.paint.ImagePattern;
-import javafx.scene.shape.Circle;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Pane;
-
-/**
- * The DialogBox Class.
- * Responsible for formatting the messages.
- *
- * @author Shishir
- */
-public class DialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private Pane messageOutline;
-    @FXML
-    private Circle circle;
-
-    /**
-     * Constructs the DialogBox.
-     * @param text User's message.
-     * @param img User's profile picture.
-     * @param isUser User's type.
-     */
-    private DialogBox(String text, Image img, boolean isUser) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        dialog.setText(text);
-        dialog.setWrapText(true);
-        dialog.setPrefWidth(260);
-        circle.setFill(new ImagePattern(img));
-        messageOutline.setPrefWidth(280);
-        messageOutline.prefHeightProperty().bind(dialog.heightProperty());
-        messageOutline.getStyleClass().clear();
-        if (isUser) {
-            messageOutline.getStyleClass().add("blue-background");
-        } else {
-            messageOutline.getStyleClass().add("gray-background");
-        }
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-    }
-
-    public static DialogBox getUserDialog(String text, Image img) {
-        return new DialogBox(text, img, true);
-    }
-
-    public static DialogBox getDukeDialog(String text, Image img) {
-        var db = new DialogBox(text, img, false);
-        db.flip();
-        return db;
-    }
-
-}
-
-
