+++ src/main/java/jarvis/commands/ExitCommand.java
@@ -1,0 +1,34 @@
+package jarvis.commands;
+
+import jarvis.storage.Storage;
+import jarvis.tasks.TaskList;
+import jarvis.ui.Ui;
+
+/**
+ * Represents the ExitCommand Class.
+ * Responsible for handling exit operations.
+ *
+ * @author Shishir
+ */
+public class ExitCommand extends Command {
+
+    /**
+     * Executes the respective command.
+     * @param tasks List of all the tasks.
+     * @param ui Ui for interacting with the user.
+     * @param storage Storage of the tasks.
+     */
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        return ui.leave();
+    }
+
+    /**
+     * Returns the exit status of the command.
+     * @return Exit status of the command.
+     */
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+
+}
