+++ src/main/java/jarvis/ui/Main.java
@@ -1,0 +1,32 @@
+package jarvis.ui;
+
+import java.io.IOException;
+
+import jarvis.Jarvis;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+
+/**
+ * A GUI for Duke using FXML.
+ */
+public class Main extends Application {
+    private Jarvis jarvis = new Jarvis();
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            stage.setResizable(false);
+            stage.setTitle("Jarvis, Your Personal Task Manager!");
+            fxmlLoader.<MainWindow>getController().setDuke(jarvis);
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
