+++ src/main/resources/css/Application.css
@@ -24,12 +24,6 @@
     -fx-border-color: "#2e333d";
 }
 
-.message-font {
-
-}
-
-.header {
-}
 
 .scroll-pane {
     -fx-background: #202329;
